Modularizado/requirements.txt
python-dotenv
pytelegrambotapi
groq
torch
certifi
charset-normalizer
colorama
filelock
fsspec
huggingface-hub
idna
Jinja2
MarkupSafe
mpmath
networkx
numpy
packaging
pillow
PyYAML
regex
requests
safetensors
setuptools
sympy
tokenizers
torchaudio
torchvision
tqdm
transformers
typing_extensions
urllib3
